requirements.txt
# 基础依赖
streamlit>=1.22.0
requests>=2.28.2
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.3

# 语音识别
vosk>=0.3.45
pyaudio>=0.2.13

# 大语言模型API
openai>=1.3.0

# 工具和辅助库
tqdm>=4.65.0
pandas>=2.0.1
matplotlib>=3.7.1

pytest==7.4.3
# 其他依赖项...